requirements.txt
transformers==4.36.2
diffusers==0.25.0
torch==2.1.0
flash_attn==2.5.8
huggingface_hub==0.23.0
safetensors==0.4.5
tokenizers==0.15.0
xformers==0.0.22
decord==0.6.0
einops==0.3.0
numpy==1.24.2
omegaconf==2.1.1
pandas==2.0.0
Pillow==9.5.0
pytorch_lightning==1.9.3
PyYAML==6.0
setuptools==65.6.3
tqdm==4.65.0
open_clip_torch==2.22.0
torchdata==0.7.0
timm
webdataset
hydra-core
pyrootutils
accelerate
scikit-learn
imageio
fairscale
av
flask
opencv_python
moviepy
kornia
